requirements.txt
# Core Dependencies
numpy>=1.24.3,<2.0.0
pandas>=2.0.3,<3.0.0
scikit-learn>=1.3.0,<2.0.0

# Deep Learning (for GRU model)
# Note: TensorFlow requires Python 3.9-3.12 (not 3.13)
# For Apple Silicon (M1/M2/M3): Use tensorflow-macos + tensorflow-metal
# For Intel Mac/Linux/Windows: Use tensorflow
tensorflow>=2.13.0,<3.0.0; platform_machine != "arm64" and python_version < "3.13"
tensorflow-macos>=2.13.0,<3.0.0; platform_machine == "arm64" and python_version < "3.13"
tensorflow-metal>=1.0.0; platform_machine == "arm64" and python_version < "3.13"

# API and WebSocket
requests>=2.31.0
websocket-client>=1.6.1

# Telegram
python-telegram-bot>=20.4

# Configuration
PyYAML>=6.0.1

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Logging
colorlog>=6.7.0

# Model Persistence
joblib>=1.3.2

# Data Processing
scipy>=1.10.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
